feat(BC-1): settings fix
