+++ apps/likes/serializers.py
@@ -1,0 +1,8 @@
+.from rest_framework import serializers
+from .models import Like
+
+
+class LikeSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Like
+        fields = "__all__"
