+++ apps/posts/models.py
@@ -1,3 +1,9 @@
 from django.db import models
 
-# Create your models here.
+
+class Post(models.Model):
+    title = models.CharField()
+    content = models.TextField()
+    user = models.ForeignKey('BlogUser')
+    timestamp = models.DateTimeField()
+    auto_now = True
