+++ apps/posts/serializers.py
@@ -1,0 +1,8 @@
+from .models import Post
+from rest_framework import serializers
+
+
+class PostSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Post
+        fields = "__all__"
