+++ apps/posts/views.py
@@ -1,3 +1,8 @@
-from django.shortcuts import render
+from rest_framework import viewsets
+from serializers import PostSerializer
+from .models import Post
 
-# Create your views here.
+
+class PostViewSet(viewsets.ModelViewSet):
+    queryset = Post.objects
+    serializer_class = PostSerializer
